+++ api/lib/tests/test_aes_xts_encrypt_decrypt_aesgenerate.rs
@@ -51,7 +51,7 @@ fn test_aes_xts_encrypt_decrypt() {
         );
         assert!(result.is_ok(), "result {:?}", result);
         let encrypted_data = result.unwrap();
-        assert_eq!(encrypted_data.data.len(), data_len as usize);
+        assert_eq!(encrypted_data.data.len(), data_len);
         assert_ne!(data, encrypted_data.data);
 
         let decrypted_result = app_session.aes_xts_encrypt_decrypt(
@@ -110,7 +110,7 @@ fn test_aes_xts_tampered_data() {
         );
         assert!(result.is_ok(), "result {:?}", result);
         let encrypted_data = result.unwrap();
-        assert_eq!(encrypted_data.data.len(), data_len as usize);
+        assert_eq!(encrypted_data.data.len(), data_len);
         assert_ne!(data, encrypted_data.data);
 
         let mut tampered_data = encrypted_data.data.clone();
@@ -180,7 +180,7 @@ fn test_aes_xts_tampered_tweak_value() {
         );
         assert!(result.is_ok(), "result {:?}", result);
         let encrypted_data = result.unwrap();
-        assert_eq!(encrypted_data.data.len(), data_len as usize);
+        assert_eq!(encrypted_data.data.len(), data_len);
         assert_ne!(data, encrypted_data.data);
 
         tweak[0] += 1;
@@ -240,7 +240,7 @@ fn test_aes_xts_encrypt_decrypt_multi_times() {
         );
         assert!(result.is_ok(), "result {:?}", result);
         let encrypted_data = result.unwrap();
-        assert_eq!(encrypted_data.data.len(), data_len as usize);
+        assert_eq!(encrypted_data.data.len(), data_len);
         assert_ne!(data, encrypted_data.data);
 
         let decrypted_result = app_session.aes_xts_encrypt_decrypt(
@@ -270,7 +270,7 @@ fn test_aes_xts_encrypt_decrypt_multi_times() {
         );
         assert!(result.is_ok(), "result {:?}", result);
         let encrypted_data = result.unwrap();
-        assert_eq!(encrypted_data.data.len(), data_len as usize);
+        assert_eq!(encrypted_data.data.len(), data_len);
         assert_ne!(data, encrypted_data.data);
 
         let decrypted_result = app_session.aes_xts_encrypt_decrypt(
@@ -330,7 +330,7 @@ fn test_aes_xts_encrypt_data_size() {
         );
         assert!(result.is_ok(), "result {:?}", result);
         let encrypted_data = result.unwrap();
-        assert_eq!(encrypted_data.data.len(), data_len as usize);
+        assert_eq!(encrypted_data.data.len(), data_len);
         assert_ne!(data, encrypted_data.data);
 
         //size 512 allowed,  smaller than data size
@@ -543,7 +543,7 @@ fn test_aes_xts_encrypt_data_size() {
         );
         assert!(result.is_ok(), "result {:?}", result);
         let encrypted_data = result.unwrap();
-        assert_eq!(encrypted_data.data.len(), data_len as usize);
+        assert_eq!(encrypted_data.data.len(), data_len);
         assert_ne!(data, encrypted_data.data);
 
         //size > 16
@@ -566,7 +566,7 @@ fn test_aes_xts_encrypt_data_size() {
         //
         assert!(result.is_ok(), "result {:?}", result);
         let encrypted_data = result.unwrap();
-        assert_eq!(encrypted_data.data.len(), data_len as usize);
+        assert_eq!(encrypted_data.data.len(), data_len);
         assert_ne!(data, encrypted_data.data);
 
         //size is odd
@@ -589,7 +589,7 @@ fn test_aes_xts_encrypt_data_size() {
         //
         assert!(result.is_ok(), "result {:?}", result);
         let encrypted_data = result.unwrap();
-        assert_eq!(encrypted_data.data.len(), data_len as usize);
+        assert_eq!(encrypted_data.data.len(), data_len);
         assert_ne!(data, encrypted_data.data);
 
         //big size
@@ -607,7 +607,7 @@ fn test_aes_xts_encrypt_data_size() {
         );
         assert!(result.is_ok(), "result {:?}", result);
         let encrypted_data = result.unwrap();
-        assert_eq!(encrypted_data.data.len(), data_len as usize);
+        assert_eq!(encrypted_data.data.len(), data_len);
         assert_ne!(data, encrypted_data.data);
     });
 }
@@ -653,7 +653,7 @@ fn test_aes_xts_decrypt_data_size() {
         );
         assert!(result.is_ok(), "result {:?}", result);
         let encrypted_data = result.unwrap();
-        assert_eq!(encrypted_data.data.len(), data_len as usize);
+        assert_eq!(encrypted_data.data.len(), data_len);
         assert_ne!(data, encrypted_data.data);
 
         // len 512 but samller than original data,
@@ -684,7 +684,7 @@ fn test_aes_xts_decrypt_data_size() {
         );
         assert!(result.is_ok(), "result {:?}", result);
         let encrypted_data = result.unwrap();
-        assert_eq!(encrypted_data.data.len(), data_len as usize);
+        assert_eq!(encrypted_data.data.len(), data_len);
         assert_ne!(data, encrypted_data.data);
 
         // dul same as buffer for 256 there is no other option
@@ -718,7 +718,7 @@ fn test_aes_xts_decrypt_data_size() {
         );
         assert!(result.is_ok(), "result {:?}", result);
         let encrypted_data = result.unwrap();
-        assert_eq!(encrypted_data.data.len(), data_len as usize);
+        assert_eq!(encrypted_data.data.len(), data_len);
         assert_ne!(data, encrypted_data.data);
 
         let decrypted_result = app_session.aes_xts_encrypt_decrypt(
@@ -748,7 +748,7 @@ fn test_aes_xts_decrypt_data_size() {
         );
         assert!(result.is_ok(), "result {:?}", result);
         let encrypted_data = result.unwrap();
-        assert_eq!(encrypted_data.data.len(), data_len as usize);
+        assert_eq!(encrypted_data.data.len(), data_len);
         assert_ne!(data, encrypted_data.data);
 
         let decrypted_result = app_session.aes_xts_encrypt_decrypt(
@@ -779,7 +779,7 @@ fn test_aes_xts_decrypt_data_size() {
         );
         assert!(result.is_ok(), "result {:?}", result);
         let encrypted_data = result.unwrap();
-        assert_eq!(encrypted_data.data.len(), data_len as usize);
+        assert_eq!(encrypted_data.data.len(), data_len);
         assert_ne!(data, encrypted_data.data);
 
         let decrypted_result = app_session.aes_xts_encrypt_decrypt(
@@ -811,7 +811,7 @@ fn test_aes_xts_decrypt_data_size() {
         );
         assert!(result.is_ok(), "result {:?}", result);
         let encrypted_data = result.unwrap();
-        assert_eq!(encrypted_data.data.len(), data_len as usize);
+        assert_eq!(encrypted_data.data.len(), data_len);
         assert_ne!(data, encrypted_data.data);
 
         let decrypted_result = app_session.aes_xts_encrypt_decrypt(
@@ -841,7 +841,7 @@ fn test_aes_xts_decrypt_data_size() {
         );
         assert!(result.is_ok(), "result {:?}", result);
         let encrypted_data = result.unwrap();
-        assert_eq!(encrypted_data.data.len(), data_len as usize);
+        assert_eq!(encrypted_data.data.len(), data_len);
         assert_ne!(data, encrypted_data.data);
 
         // len 2056, not allowed size, not the data len either
@@ -1061,7 +1061,7 @@ fn test_aes_xts_decrypt_mismatched_datalen() {
         );
         assert!(result.is_ok(), "result {:?}", result);
         let encrypted_data = result.unwrap();
-        assert_eq!(encrypted_data.data.len(), data_len as usize);
+        assert_eq!(encrypted_data.data.len(), data_len);
         assert_ne!(data, encrypted_data.data);
 
         //len < data size
@@ -1137,7 +1137,7 @@ fn test_aes_xts_switch_keys() {
         );
         assert!(result.is_ok(), "result {:?}", result);
         let encrypted_data = result.unwrap();
-        assert_eq!(encrypted_data.data.len(), data_len as usize);
+        assert_eq!(encrypted_data.data.len(), data_len);
         assert_ne!(data, encrypted_data.data);
 
         let decrypted_result = app_session.aes_xts_encrypt_decrypt(
